static typing
